backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "MediGuard — AI Health Support for Elderly in Taiwan"
requires-python = ">=3.12"
dependencies = [
    "uvicorn[standard]>=0.46.0",
    "python-dotenv>=1.2.2",
    "pydantic[email]>=2.13.3",
    "pydantic-settings>=2.14.0",
    "httpx>=0.28.1",
    "supabase>=2.29.0",
    "apscheduler>=3.10.4",
    "python-multipart>=0.0.9",
    "pyjwt>=2.12.1",
    "fastapi>=0.136.1",
    "pytest>=8.4.2",
]
